move debouncing to parent
